requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-core
langchain-chroma
pypdf2
openai
python-dotenv
tiktoken
